Struct near_sdk::collections::LazyOption

pub struct LazyOption<T> { /* fields omitted */ }

An persistent lazy option, that stores a value in the storage.

Implementations

impl<T> LazyOption<T>

pub fn is_some(&self) -> bool

Returns true if the value is present in the storage.

pub fn is_none(&self) -> bool

Returns true if the value is not present in the storage.

impl<T> LazyOption<T> where
T: BorshSerialize + BorshDeserialize,

pub fn new<S>(storage_key: S, value: Option<&T>) -> Self where
S: IntoStorageKey,

Create a new lazy option with the given storage_key and the initial value.

pub fn remove(&mut self) -> bool

Removes the value from storage without reading it. Returns whether the value was present.

pub fn take(&mut self) -> Option<T>

Removes the value from storage and returns it as an option.

pub fn get(&self) -> Option<T>

Gets the value from storage and returns it as an option.

pub fn set(&mut self, value: &T) -> bool

Sets the value into the storage without reading the previous value and returns whether the previous value was present.

pub fn replace(&mut self, value: &T) -> Option<T>

Replaces the value in the storage and returns the previous value as an option.