Struct near_sdk::collections::LookupSet

pub struct LookupSet<T> { /* fields omitted */ }

An non-iterable implementation of a set that stores its content directly on the trie.

Implementations

impl<T> LookupSet<T>

pub fn new<S>(element_prefix: S) -> Self where
S: IntoStorageKey,

Create a new map. Use element_prefix as a unique prefix for trie keys.

pub fn insert_raw(&mut self, element_raw: &[u8]) -> bool

Inserts a serialized element into the set. If the set did not have this value present, true is returned. If the set did have this value present, false is returned.

pub fn remove_raw(&mut self, element_raw: &[u8]) -> bool

Removes a serialized element from the set. Returns true if the element was present in the set.

impl<T> LookupSet<T> where
T: BorshSerialize,

pub fn contains(&self, element: &T) -> bool

Returns true if the set contains an element.

pub fn remove(&mut self, element: &T) -> bool

Removes a value from the set. Returns whether the value was present in the set.

pub fn insert(&mut self, element: &T) -> bool

Adds a value to the set. If the set did not have this value present, true is returned. If the set did have this value present, false is returned.

pub fn extend<IT: IntoIterator<Item = T>>(&mut self, iter: IT)