Struct near_sdk::collections::TreeMap

pub struct TreeMap<K, V> { /* fields omitted */ }

TreeMap based on AVL-tree

Runtime complexity (worst case):

get/contains_key: O(1) - UnorderedMap lookup

insert/remove: O(log(N))

min/max: O(log(N))

above/below: O(log(N))

range of K elements: O(Klog(N))

Implementations

impl<K, V> TreeMap<K, V> where
K: Ord + Clone + BorshSerialize + BorshDeserialize,
V: BorshSerialize + BorshDeserialize,

pub fn new<S>(prefix: S) -> Self where
S: IntoStorageKey,

pub fn len(&self) -> u64

pub fn clear(&mut self)

pub fn contains_key(&self, key: &K) -> bool

pub fn get(&self, key: &K) -> Option<V>

pub fn insert(&mut self, key: &K, val: &V) -> Option<V>

pub fn remove(&mut self, key: &K) -> Option<V>

pub fn min(&self) -> Option<K>

Returns the smallest stored key from the tree

pub fn max(&self) -> Option<K>

Returns the largest stored key from the tree

pub fn higher(&self, key: &K) -> Option<K>

Returns the smallest key that is strictly greater than key given as the parameter

pub fn lower(&self, key: &K) -> Option<K>

Returns the largest key that is strictly less than key given as the parameter

pub fn ceil_key(&self, key: &K) -> Option<K>

Returns the smallest key that is greater or equal to key given as the parameter

pub fn floor_key(&self, key: &K) -> Option<K>

Returns the largest key that is less or equal to key given as the parameter

pub fn iter<'a>(&'a self) -> impl Iterator<Item = (K, V)> + 'a

Iterate all entries in ascending order: min to max, both inclusive

pub fn iter_from<'a>(&'a self, key: K) -> impl Iterator<Item = (K, V)> + 'a

Iterate entries in ascending order: given key (exclusive) to max (inclusive)

pub fn iter_rev<'a>(&'a self) -> impl Iterator<Item = (K, V)> + 'a

Iterate all entries in descending order: max to min, both inclusive

pub fn iter_rev_from<'a>(&'a self, key: K) -> impl Iterator<Item = (K, V)> + 'a

Iterate entries in descending order: given key (exclusive) to min (inclusive)

pub fn range<'a>(
&'a self,
r: (Bound<K>, Bound<K>)
) -> impl Iterator<Item = (K, V)> + 'a

Iterate entries in ascending order according to specified bounds.

Panics

Panics if range start > end. Panics if range start == end and both bounds are Excluded.

pub fn to_vec(&self) -> Vec<(K, V)>