Struct near_sdk::collections::UnorderedSet

pub struct UnorderedSet<T> { /* private fields */ }

An iterable implementation of a set that stores its content directly on the trie.

Implementations

impl<T> UnorderedSet<T>

pub fn len(&self) -> u64

Returns the number of elements in the set, also referred to as its size.

Examples

use near_sdk::collections::UnorderedSet;

let mut set: UnorderedSet<u8> = UnorderedSet::new(b"s");
assert_eq!(set.len(), 0);
set.insert(&1);
set.insert(&2);
assert_eq!(set.len(), 2);

pub fn is_empty(&self) -> bool

Returns true if the set contains no elements.

pub fn new<S>(prefix: S) -> Self where
S: IntoStorageKey,

Create new map with zero elements. Use id as a unique identifier.

Examples

use near_sdk::collections::UnorderedSet;
let mut set: UnorderedSet<u32> = UnorderedSet::new(b"s");

pub fn insert_raw(&mut self, element_raw: &[u8]) -> bool

Adds a value to the set. If the set did not have this value present, true is returned. If the set did have this value present, false is returned.

pub fn remove_raw(&mut self, element_raw: &[u8]) -> bool

Removes a value from the set. Returns whether the value was present in the set.

impl<T> UnorderedSet<T> where
T: BorshSerialize + BorshDeserialize,

pub fn contains(&self, element: &T) -> bool

Returns true if the set contains an element.

Examples

use near_sdk::collections::UnorderedSet;

let mut set: UnorderedSet<u8> = UnorderedSet::new(b"s");
assert_eq!(set.contains(&1), false);
set.insert(&1);
assert_eq!(set.contains(&1), true);

pub fn remove(&mut self, element: &T) -> bool

Removes a value from the set. Returns whether the value was present in the set.

Examples

use near_sdk::collections::UnorderedSet;

let mut set: UnorderedSet<u8> = UnorderedSet::new(b"s");
assert_eq!(set.remove(&1), false);
set.insert(&1);
assert_eq!(set.remove(&1), true);
assert_eq!(set.contains(&1), false);

pub fn insert(&mut self, element: &T) -> bool

Adds a value to the set. If the set did not have this value present, true is returned. If the set did have this value present, false is returned.

Examples

use near_sdk::collections::UnorderedSet;

let mut set: UnorderedSet<u8> = UnorderedSet::new(b"s");
assert_eq!(set.insert(&1), true);
assert_eq!(set.insert(&1), false);
assert_eq!(set.contains(&1), true);

pub fn clear(&mut self)

Clears the map, removing all elements.

Examples

use near_sdk::collections::UnorderedSet;

let mut set: UnorderedSet<u8> = UnorderedSet::new(b"s");
set.insert(&1);
set.insert(&2);
set.clear();
assert_eq!(set.len(), 0);

pub fn to_vec(&self) -> Vec<T>

Copies elements into an std::vec::Vec.

pub fn iter(&self) -> impl Iterator<Item = T> + '_

Iterate over deserialized elements.

pub fn extend<IT: IntoIterator<Item = T>>(&mut self, iter: IT)

pub fn as_vector(&self) -> &Vector<T>

Returns a view of elements as a vector. It’s sometimes useful to have random access to the elements.