Struct near_sdk::collections::unordered_map::UnorderedMap

pub struct UnorderedMap<K, V> { /* private fields */ }

An iterable implementation of a map that stores its content directly on the trie.

Implementations

impl<K, V> UnorderedMap<K, V>

pub fn len(&self) -> u64

Returns the number of elements in the map, also referred to as its size.

Examples

use near_sdk::collections::UnorderedMap;

let mut map: UnorderedMap<u8, u8> = UnorderedMap::new(b"m");
assert_eq!(map.len(), 0);
map.insert(&1, &1);
map.insert(&2, &2);
assert_eq!(map.len(), 2);

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

pub fn new<S>(prefix: S) -> Self where
S: IntoStorageKey,

Create new map with zero elements. Use prefix as a unique identifier.

Examples

use near_sdk::collections::UnorderedMap;
let mut map: UnorderedMap<u8, u8> = UnorderedMap::new(b"m");

pub fn insert_raw(&mut self, key_raw: &[u8], value_raw: &[u8]) -> Option<Vec<u8>>

Inserts a serialized key-value pair into the map. If the map did not have this key present, None is returned. Otherwise returns a serialized value. Note, the keys that have the same hash value are undistinguished by the implementation.

pub fn remove_raw(&mut self, key_raw: &[u8]) -> Option<Vec<u8>>

Removes a serialized key from the map, returning the serialized value at the key if the key was previously in the map.

impl<K, V> UnorderedMap<K, V> where
K: BorshSerialize + BorshDeserialize,
V: BorshSerialize + BorshDeserialize,

pub fn get(&self, key: &K) -> Option<V>

Returns the value corresponding to the key.

Examples

use near_sdk::collections::UnorderedMap;

let mut map: UnorderedMap<u8, u8> = UnorderedMap::new(b"m");
assert_eq!(map.get(&1), None);
map.insert(&1, &10);
assert_eq!(map.get(&1), Some(10));

pub fn remove(&mut self, key: &K) -> Option<V>

Removes a key from the map, returning the value at the key if the key was previously in the map.

Examples

use near_sdk::collections::UnorderedMap;

let mut map: UnorderedMap<u8, u8> = UnorderedMap::new(b"m");
assert_eq!(map.remove(&1), None);
map.insert(&1, &10);
assert_eq!(map.remove(&1), Some(10));
assert_eq!(map.len(), 0);

pub fn insert(&mut self, key: &K, value: &V) -> Option<V>

Inserts a key-value pair into the map. If the map did not have this key present, None is returned. Otherwise returns a value. Note, the keys that have the same hash value are undistinguished by the implementation.

Examples

use near_sdk::collections::UnorderedMap;

let mut map: UnorderedMap<u8, u8> = UnorderedMap::new(b"m");
map.insert(&1, &10);
assert_eq!(map.get(&1), Some(10));
assert_eq!(map.len(), 1);

pub fn clear(&mut self)

Clears the map, removing all elements.

Examples

use near_sdk::collections::UnorderedMap;

let mut map: UnorderedMap<u8, u8> = UnorderedMap::new(b"m");
map.insert(&1, &10);
map.insert(&2, &20);
map.clear();
assert_eq!(map.len(), 0);

pub fn to_vec(&self) -> Vec<(K, V)>

Copies elements into an std::vec::Vec.

pub fn keys(&self) -> impl Iterator<Item = K> + '_

An iterator visiting all keys. The iterator element type is K.

pub fn values(&self) -> impl Iterator<Item = V> + '_

An iterator visiting all values. The iterator element type is V.

pub fn iter(&self) -> Iter<'_, K, V>

Iterate over deserialized keys and values.

pub fn extend<IT: IntoIterator<Item = (K, V)>>(&mut self, iter: IT)

pub fn keys_as_vector(&self) -> &Vector<K>

Returns a view of keys as a vector. It’s sometimes useful to have random access to the keys.

pub fn values_as_vector(&self) -> &Vector<V>

Returns a view of values as a vector. It’s sometimes useful to have random access to the values.