Enum near_sdk::PromiseOrValue

pub enum PromiseOrValue<T> {
    Promise(Promise),
    Value(T),
}

When the method can return either a promise or a value, it can be called with PromiseOrValue::Promise or PromiseOrValue::Value to specify which one should be returned.

Example

#[ext_contract]
pub trait ContractA {
    fn a(&mut self);
}

let value = Some(true);
let val: PromiseOrValue<bool> = if let Some(value) = value {
    PromiseOrValue::Value(value)
} else {
    contract_a::ext("bob_near".parse().unwrap()).a().into()
};

Variants

Promise(Promise)

Value(T)