Module near_sdk::json_types

Helper types for JSON serialization.

Structs

Base58CryptoHash

Base58PublicKey
Public key in a binary format with base58 string serialization with human-readable curve. e.g. ed25519:3tysLvy7KGoE8pznUgXvSHa4vYyGvrDZFcT8jgb8PEQ6

Base64VecU8
Helper class to serialize/deserialize Vec<u8> to base64 string.

I64

I128

U64

U128

ValidAccountId
Helper class to validate account ID during serialization and deserializiation

Enums

CurveType
PublicKey curve

Type Definitions

WrappedBalance
Balance wrapped into a struct for JSON serialization as a string.

WrappedDuration
Duration in nanosecond wrapped into a struct for JSON serialization as a string.

WrappedTimestamp
Timestamp in nanosecond wrapped into a struct for JSON serialization as a string.