Struct near_sdk::AccountId

pub struct AccountId(_);

Account identifier. This is the human readable utf8 string which is used internally to index accounts on the network and their respective state.

Because these IDs have to be validated, they have to be converted from a string with FromStr or TryFrom a compatible type. To skip validation on initialization, AccountId::new_unchecked can be used.

Examples

use near_sdk::AccountId;
use std::convert::{TryFrom, TryInto};

// `FromStr` conversion
let alice: AccountId = "alice.near".parse().unwrap();
assert!("invalid.".parse::<AccountId>().is_err());

let alice_string = "alice".to_string();

// From string with validation
let alice = AccountId::try_from(alice_string.clone()).unwrap();
let alice: AccountId = alice_string.try_into().unwrap();

// Initialize without validating
let alice_unchecked = AccountId::new_unchecked("alice".to_string());
assert_eq!(alice, alice_unchecked);

Implementations

impl AccountId

pub fn as_bytes(&self) -> &[u8]

Returns reference to the account ID bytes.

pub fn as_str(&self) -> &str

Returns reference to the account ID string.

pub fn new_unchecked(id: String) -> Self

Caller must ensure that the account id is valid.

For more information, read: https://docs.near.org/docs/concepts/account#account-id-rules