Struct near_sdk::MockedBlockchain

pub struct MockedBlockchain { /* fields omitted */ }

Mocked blockchain that can be used in the tests for the smart contracts. It implements BlockchainInterface by redirecting calls to VMLogic. It unwraps errors of VMLogic to cause panic during the unit tests similarly to how errors of VMLogic would cause the termination of guest program execution. Unit tests can even assert the expected error message.

Implementations

impl MockedBlockchain

pub fn new(
context: VMContext,
config: VMConfig,
fees_config: RuntimeFeesConfig,
promise_results: Vec<PromiseResult>,
storage: HashMap<Vec<u8>, Vec<u8>>,
validators: HashMap<AccountId, Balance>,
memory_opt: Option<Box<dyn MemoryLike>>
) -> Self

pub fn take_storage(&mut self) -> HashMap<Vec<u8>, Vec<u8>>

pub fn created_receipts(&self) -> &Vec<Receipt>

pub fn outcome(&self) -> VMOutcome

pub fn gas(&mut self, gas_amount: u32)

pub fn logs(&self) -> Vec<String>