Module near_sdk::test_utils

Testing blockchain utilities. These can only be used inside tests and are not available for a wasm32 target.

Modules

test_env

Structs

VMContextBuilder

Simple VMContext builder that allows to quickly create custom context in tests.

Functions

accounts

Returns a pre-defined account_id from a list of 6.

get_created_receipts

Accessing receipts created by the contract. Only available in unit tests.

get_logs

Returns a copy of logs from VMLogic. Only available in unit tests.

testing_env_with_promise_resultsDeprecated

Initializes the MockedBlockchain with a single promise result during execution.