Module near_sdk::test_utils::test_env

Functions

alice

bob

carol

setupDeprecated

Setup the blockchain interface with a default configuration.

setup_free

free == effectively unlimited gas Sets up the blockchain interface with a VMConfig which sets the gas costs to zero.

setup_with_configDeprecated

Updates the blockchain interface with the config passed in.