Module near_sdk::utils

Structs

PendingContractTx
Used in the simulation code generator from near_sdk.

Functions

assert_one_yocto
Assert that 1 yoctoNEAR was attached.

assert_self
Assert that predecessor_account_id == current_account_id, meaning contract called itself.

is_promise_success
Returns true if promise was successful. Fails if called outside a callback that received 1 promise result.

promise_result_as_success
Returns the result of the promise if successful. Otherwise returns None. Fails if called outside a callback that received 1 promise result.