Derive Macro near_sdk::EventMetadata

#[derive(EventMetadata)]
{
    // Attributes available to this derive:
    #[event_version]
}

NOTE: This is an internal implementation for #[near_bindgen(events(standard = ...))] attribute.

This derive macro is used to inject the necessary wrapper and logic to auto format standard event logs. The other appropriate attribute macros are not injected with this macro. Required attributes below:

#[derive(near_sdk::serde::Serialize, std::clone::Clone)]
#[serde(crate="near_sdk::serde")]
#[serde(tag = "event", content = "data")]
#[serde(rename_all="snake_case")]
pub enum MyEvent {
    Event
}